Introduce constant TB_OBJECT_MODELS_AUTHORITATIVE
This constants exists to tell CoreUpdate to consider object models definitions
as an ultimate source of truth. Database tables will be migrated to match
these definitions if there any differences in table structure are found.

Because this constant didn't exist in older version of thirtybess, using core updater
to downgrade to older version will not trigger any database migration